Force color counter to reset at end of color list.
Resetting color counter to avoid out of range errors. Now the code
will keep cycling through the list it was given until it runs out
of new features.

Made a few documentation updates that were missed in the previous
commit.

Removed (commented out) old color and finish functionality that
survived the purge in the previous commit.